refactor: Split Util to several files
