+++ src/Util/QCodeEditorUtil.cpp
@@ -1,0 +1,76 @@
+/*
+ * Copyright (C) 2019-2020 Ashar Khan <ashar786khan@gmail.com>
+ *
+ * This file is part of CP Editor.
+ *
+ * CP Editor is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * I will not be responsible if CP Editor behaves in unexpected way and
+ * causes your ratings to go down and or lose any important contest.
+ *
+ * Believe Software is "Software" and it isn't immune to bugs.
+ *
+ */
+
+#include "Util/QCodeEditorUtil.hpp"
+#include "Core/EventLogger.hpp"
+#include "Extensions/EditorTheme.hpp"
+#include <QCXXHighlighter>
+#include <QCodeEditor>
+#include <QJavaHighlighter>
+#include <QPythonHighlighter>
+#include <generated/SettingsHelper.hpp>
+
+namespace Util
+{
+void applySettingsToEditor(QCodeEditor *editor)
+{
+    LOG_INFO("Applying settings to QCodeEditor");
+
+    editor->setTabReplace(SettingsHelper::isReplaceTabs());
+    editor->setTabReplaceSize(SettingsHelper::getTabWidth());
+    editor->setAutoIndentation(SettingsHelper::isAutoIndent());
+    editor->setAutoParentheses(SettingsHelper::isAutoCompleteParentheses());
+    editor->setAutoRemoveParentheses(SettingsHelper::isAutoRemoveParentheses());
+
+    editor->setFont(SettingsHelper::getEditorFont());
+
+    const int tabStop = SettingsHelper::getTabWidth();
+    QFontMetrics metric(editor->font());
+    editor->setTabReplaceSize(tabStop);
+
+    if (SettingsHelper::isWrapText())
+        editor->setWordWrapMode(QTextOption::WordWrap);
+    else
+        editor->setWordWrapMode(QTextOption::NoWrap);
+
+    auto style = Extensions::EditorTheme::query(SettingsHelper::getEditorTheme());
+    if (!style)
+        style = Extensions::EditorTheme::query("Light");
+    editor->setSyntaxStyle(style);
+
+    editor->setExtraBottomMargin(SettingsHelper::isExtraBottomMargin());
+}
+
+void setEditorLanguage(QCodeEditor *editor, const QString &language)
+{
+    if (language == "Python")
+    {
+        editor->setHighlighter(new QPythonHighlighter);
+        editor->setCompleter(nullptr);
+    }
+    else if (language == "Java")
+    {
+        editor->setHighlighter(new QJavaHighlighter);
+        editor->setCompleter(nullptr);
+    }
+    else
+    {
+        editor->setHighlighter(new QCXXHighlighter);
+        editor->setCompleter(nullptr);
+    }
+}
+} // namespace Util
