+++ src/Util/README.md
@@ -1,0 +1,3 @@
+# Util
+
+This directory contains functions and constants which are useful in many places.
