+++ src/main.cpp
@@ -27,6 +27,7 @@
 #include <QJsonObject>
 #include <QTextStream>
 #include <generated/version.hpp>
+#include <iostream>
 #include <singleapplication.h>
 
 #define TOJSON(x) json[#x] = x
